requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Werkzeug
marshmallow
pyodbc
Flask-cors
Flask-restful
Python-dotenv
Flask-sqlalchemy